refactor(web): auth 인증 entities 분리
