+++ apps/web/src/features/analysis/api/index.ts
@@ -1,5 +1,5 @@
+import { createAuthedRestAPI } from "@/entities/auth/lib/create-authed-rest";
 import { AnalysisAPIService } from "@/features/analysis/api/analysis-api.service";
-import { createAuthedRestAPI } from "@/features/settings/lib/create-authed-rest";
 
 export const analysisAPIService = new AnalysisAPIService(
   createAuthedRestAPI(process.env.NEXT_PUBLIC_BACKEND_URL || ""),
