+++ apps/web/src/features/settings/api/index.ts
@@ -1,23 +1,5 @@
-import { generateRestAPI } from "@recap/api";
-
-import { AuthAPIService } from "@/features/settings/api/auth-api.service";
+import { createAuthedRestAPI } from "@/entities/auth/lib/create-authed-rest";
 import { UserAPIService } from "@/features/settings/api/user-api.service";
-import { createAuthedRestAPI } from "@/features/settings/lib/create-authed-rest";
-
-export const authAPIService = new AuthAPIService(
-  generateRestAPI(
-    {
-      APIbaseURL: `api/v1/auth`,
-    },
-    {
-      baseURL: process.env.NEXT_PUBLIC_BACKEND_URL || "",
-    },
-  ),
-);
-
-export const authWithTokenAPIService = new AuthAPIService(
-  createAuthedRestAPI(process.env.NEXT_PUBLIC_BACKEND_URL || ""),
-);
 
 export const userAPIService = new UserAPIService(
   createAuthedRestAPI(process.env.NEXT_PUBLIC_BACKEND_URL || ""),
