+++ apps/web/src/pages/ai-recap/ui/AiRecapPage.tsx
@@ -5,12 +5,12 @@ import Image, { type StaticImageData } from "next/image";
 import { useQuery } from "@tanstack/react-query";
 import { useQueryClient } from "@tanstack/react-query";
 
+import { useAuthStatus } from "@/entities/auth/model/use-auth-status";
 import { recapAPIService } from "@/features/ai-recap/api";
 import type { NormalizedRecap } from "@/features/ai-recap/model/recap.type";
 import RecapSummary from "@/features/ai-recap/ui/RecapSummary";
 import Timeline from "@/features/ai-recap/ui/Timeline";
 import TopVisitedTopics from "@/features/ai-recap/ui/TopVisitedTopics";
-import { useAuthStatus } from "@/features/settings/lib/use-auth-status";
 import EmptyRecapImg1 from "@/shared/assets/img/empty-reacp-1.png";
 import EmptyRecapImg2 from "@/shared/assets/img/empty-recap-2.png";
 import LoginBanner from "@/shared/ui/LoginBanner";
